+++ pbm/utils.py
@@ -28,3 +28,17 @@ def json_default(o):
     else:
         return o
 
+
+import jinja2
+import os.path
+
+
+def get_template(template):
+    env = jinja2.Environment(autoescape=True)
+    loader = jinja2.FileSystemLoader(
+        os.path.join(
+            os.path.dirname(
+                os.path.abspath(__file__)),
+            'templates'))
+    tmpl = loader.load(env, template)
+    return tmpl